MSVC and Clang warning fixes.

+++ lib/goom/include/goom/goom_graphic.h
@@ -113,7 +113,7 @@ inline const Pixel Pixel::WHITE{/*.channels*/ {/*.r = */ MAX_COLOR_VAL, /*.g = *
 
 struct FXBuffSettings
 {
-  static constexpr float INITIAL_BUFF_INTENSITY = 0.5;
+  static constexpr float INITIAL_BUFF_INTENSITY = 0.5F;
   float buffIntensity = INITIAL_BUFF_INTENSITY;
   bool allowOverexposed = true;
 };
@@ -306,14 +306,16 @@ inline auto PixelBuffer::GetRowIter(const size_t y)
     -> std::tuple<PixelBuffer::iterator, PixelBuffer::iterator>
 {
   const auto rowPos = static_cast<int32_t>(y * m_width);
-  return std::make_tuple(m_buff.begin() + rowPos, m_buff.begin() + rowPos + m_width);
+  return std::make_tuple(m_buff.begin() + rowPos,
+                         m_buff.begin() + rowPos + static_cast<int32_t>(m_width));
 }
 
 inline auto PixelBuffer::GetRowIter(const size_t y) const
     -> std::tuple<PixelBuffer::const_iterator, PixelBuffer::const_iterator>
 {
   const auto rowPos = static_cast<int32_t>(y * m_width);
-  return std::make_tuple(m_buff.begin() + rowPos, m_buff.begin() + rowPos + m_width);
+  return std::make_tuple(m_buff.begin() + rowPos,
+                         m_buff.begin() + rowPos + static_cast<int32_t>(m_width));
 }
 
 inline auto PixelBuffer::Get4RHBNeighbours(const size_t x, const size_t y) const
